adjusted predator AC to match new health slider values

+++ Assets/Scripts/Steven/DeathReason.cs
@@ -3,5 +3,5 @@ using UnityEngine;
 public class DeathReason : MonoBehaviour
 {
     public string reason;
-    public int ac;
+    public float ac;
 }
